# openapi.yaml
openapi: 3.1.0
info:
  title: Fluxora HTTP API
  version: 0.1.0
  description: |
    Fluxora Backend: Off-chain companion to the streaming contract.

    Presents operator-grade reliability with predictable HTTP semantics, durable views of chain-derived data,
    and explicit failure behavior when dependencies are unhealthy.

    ## Security Header Baseline
    Fluxora applies Helmet at the Express boundary so every HTTP response carries a consistent
    browser-facing security-header baseline, including `Content-Security-Policy`,
    `Strict-Transport-Security`, `X-Content-Type-Options`, `X-Frame-Options`, and
    `Referrer-Policy`. The service also suppresses the `X-Powered-By` header.

    These headers are transport invariants for public clients, authenticated partners,
    administrators, and internal workers that traverse the HTTP surface. They do not replace
    authentication, authorization, TLS termination, or proxy hardening.

    ## Trust Boundaries
    - **Public Internet Clients**: Read-only access to stream listings and health checks
    - **Authenticated Partners**: Create and manage streams with idempotency guarantees
    - **Administrators**: Full access including internal indexer endpoints
    - **Internal Workers**: Indexer endpoints for chain state synchronization

    ## Failure Modes & Client-Visible Behavior
    - **Invalid Input (400)**: Malformed addresses, negative amounts, oversized payloads
...

  contact:
    name: Fluxora Team
  license:
...
servers:
...
paths:
...
  /api/streams:
    post:
      summary: Create a new stream
      operationId: createStream
      tags:
        - Streams
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: |
            Unique key that makes this request idempotent. Clients MUST supply a
            fresh value per logical operation and MUST reuse the same value when
            retrying after a network failure.

            **Format**: 1–128 characters; allowed charset: `[A-Za-z0-9:_-]`.
            UUID v4 is recommended but any value matching the charset is accepted.

            **Behaviour**:
            - Same key + same body → 201 (cached response, `Idempotency-Replayed: true`)
            - Same key + different body → 409 CONFLICT
            - Missing or malformed key → 400 VALIDATION_ERROR

            **Security**: The key value is never echoed in error response bodies
            or server logs to prevent accidental secret leakage.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStreamRequest'
      responses:
        '201':
          description: |
            Stream created (or replayed from cache).
            Check `Idempotency-Replayed` response header and `meta.idempotencyReplayed`
            to distinguish a fresh creation from a cached replay.
          headers:
            Idempotency-Key:
              description: Echoes the key supplied by the client.
              schema:
                type: string
            Idempotency-Replayed:
              description: |
                `true` when the response was served from the idempotency cache;
                `false` on first creation.
              schema:
                type: string
                enum: ['true', 'false']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamResponse'
        '400':
          description: |
            Invalid request. Common causes:
            - Missing or malformed `Idempotency-Key` header
            - Invalid JSON body
            - Invalid Stellar address format
            - Non-string amount field (must be decimal string)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            Idempotency-Key collision — the same key was previously used with a
            different request body. Use a new Idempotency-Key or retry with the
            original body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds 256 KiB limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Unprocessable entity — one or more Stellar addresses in the request
            body do not exist on-chain. The stream cannot be created because it
            would never match an on-chain event.

            The `details.missingAddresses` array lists every address that was
            not found. Addresses that could not be verified due to an RPC outage
            are omitted (the request is allowed through in that case).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - type: object
                    properties:
                      error:
                        type: object
                        properties:
                          details:
                            type: object
                            properties:
                              missingAddresses:
                                type: array
                                items:
                                  type: string
                                  pattern: '^G[A-Z2-7]{55}$'
                                example:
                                  - 'GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN'
        '503':
          description: Service unavailable (database, Stellar RPC, or idempotency store down)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List streams with cursor pagination and indexed filters
      operationId: listStreams
      tags:
        - Streams
      parameters:
        - name: limit
          in: query
          required: false
          description: Number of streams to return (1-100, default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: |
            Opaque pagination cursor returned in `next_cursor` from a previous
            response. Encodes the last seen stream id (base64url). Omit for
            the first page.
          schema:
            type: string
        - name: include_total
          in: query
          required: false
          description: |
            When true, include a `total` field in the response reflecting the
            current count of matching streams. This is a point-in-time snapshot
            and is not guaranteed to be consistent with the cursor position.
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          required: false
          description: Filter by stream status (uses the status index).
          schema:
            type: string
            enum: [scheduled, active, paused, completed, cancelled]
        - name: sender
          in: query
          required: false
          description: Filter by sender Stellar address (uses the sender_address index).
          schema:
            type: string
        - name: recipient
          in: query
          required: false
          description: Filter by recipient Stellar address (uses the recipient_address index).
          schema:
            type: string
      responses:
        '200':
          description: |
            Cursor-paginated list of streams.

            **Ordering guarantee**: results are always returned in `created_at DESC, id DESC`
            order. Because `id` is the primary key (unique), the composite key is unique and
            the ordering is deterministic — even when multiple streams share the same
            `created_at` timestamp (e.g. batch inserts). Clients can rely on a stable,
            consistent page ordering across all requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/streams/{streamId}:
    get:
      summary: Get stream details
      operationId: getStream
      tags:
        - Streams
      parameters:
        - name: streamId
          in: path
          required: true
          description: 'Stream identifier (format: stream-{timestamp}-{random})'
          schema:
            type: string
      responses:
        '200':
          description: Stream details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamResponse'
        '404':
          description: Stream not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - timestamp
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp of the response
          example: '2024-01-01T12:00:00.000Z'
        requestId:
          type: string
          description: Opaque request identifier for log correlation
          example: 550e8400-e29b-41d4-a716-446655440000
...
    CreateStreamRequest:
      type: object
      required:
        - sender
        - recipient
        - depositAmount
        - ratePerSecond
        - startTime
      properties:
        sender:
          type: string
          description: |
            Stellar public key of the sender.
            Must start with 'G' followed by exactly 55 base32 characters [A-Z2-7].
            Total length: 56 characters.
          pattern: '^G[A-Z2-7]{55}$'
          example: GBBD47UZQ5CYVVEUVRYNQZX3G5KRZTAYF5XSVS2UKMCCWW5LJJLXNVQX
        recipient:
          type: string
          description: |
            Stellar public key of the recipient.
            Must start with 'G' followed by exactly 55 base32 characters [A-Z2-7].
            Total length: 56 characters.
          pattern: '^G[A-Z2-7]{55}$'
          example: GBRPYHIL2CI3WHZDTOOQFC6EB4KJJGUJJBBX7XNLG5DBNVQWDADUZSQX
        depositAmount:
          type: string
          description: |
            Total amount to deposit in stroops (1 XLM = 10,000,000 stroops).
            Must be a non-negative decimal string.
            Must be >= ratePerSecond (minimum 1 second of streaming).
            Max: 9223372036854775807 stroops (~922 million XLM).
          pattern: '^\d+(\.\d+)?$'
          example: '1000000000'
        ratePerSecond:
          type: string
          description: |
            Streaming rate in stroops per second.
            Must be a positive decimal string.
          pattern: '^\d+(\.\d+)?$'
          example: '100000'
        startTime:
          type: integer
          description: |
            Unix timestamp (seconds) when streaming begins.
            Must be in the future or within the last hour.
          example: 1704067200
    Stream:
      type: object
      description: |
        A durable, PostgreSQL-backed stream record.
        All monetary amounts are decimal strings - never numbers.
      required:
        - id
        - sender
        - recipient
        - depositAmount
        - ratePerSecond
        - startTime
        - endTime
        - status
      properties:
        id:
          type: string
          description: |
            Unique stream identifier derived from SHA-256 of request content.
            Format: stream-{64-char-hex}-{eventIndex}
          example: stream-a3f1c2d...0c-0
        sender:
          type: string
          description: |
            Sender Stellar public key (G + 55 base32 chars).
          pattern: '^G[A-Z2-7]{55}$'
          example: GBBD47UZQ5CYVVEUVRYNQZX3G5KRZTAYF5XSVS2UKMCCWW5LJJLXNVQX
        recipient:
          type: string
          description: |
            Recipient Stellar public key (G + 55 base32 chars).
          pattern: '^G[A-Z2-7]{55}$'
          example: GBRPYHIL2CI3WHZDTOOQFC6EB4KJJGUJJBBX7XNLG5DBNVQWDADUZSQX
        depositAmount:
          type: string
          description: Total deposit as a decimal string (never a number).
          pattern: '^[+-]?[0-9]+(\.[0-9]+)?$'
          example: '1000'
        ratePerSecond:
          type: string
          description: Streaming rate per second as a decimal string.
          pattern: '^[+-]?[0-9]+(\.[0-9]+)?$'
          example: '10'
        startTime:
          type: integer
          description: Stream start time (Unix epoch seconds).
          example: 1700000000
        endTime:
          type: integer
          description: Stream end time (Unix epoch seconds). 0 = indefinite.
          example: 0
        status:
          type: string
          enum:
            - active
            - paused
            - completed
            - cancelled
          description: Current stream status.
          example: active
    StreamResponse:
      type: object
      required:
        - streams
        - has_more
        - next_cursor
      properties:
        streams:
          type: array
          items:
            $ref: '#/components/schemas/Stream'
        has_more:
          type: boolean
          description: Whether more pages are available.
          example: true
        next_cursor:
          type: string
          nullable: true
          description: |
            Opaque cursor for the next page. Always present in the response.
            `null` when `has_more` is false (no further pages).
            Pass as the `cursor` query parameter on the next request.
          example: eyJ2IjoxLCJsYXN0SWQiOiJzdHJlYW0tYWJjLTAifQ
        total:
          type: integer
          nullable: true
          description: Total count (only present when include_total=true).
          example: 42
    StreamListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required:
            - streams
            - has_more
            - next_cursor
          properties:
            streams:
              type: array
              items:
                $ref: '#/components/schemas/Stream'
            has_more:
              type: boolean
              description: Whether more streams are available.
              example: false
            next_cursor:
              type: string
              nullable: true
              description: |
                Opaque cursor for the next page. Always present.
                `null` when there are no further pages.
                Encodes the `id` of the last stream on the current page
                using keyset (`WHERE id > $cursor ORDER BY id ASC`) semantics.
              example: null
            total:
              type: integer
              nullable: true
              description: Total number of streams (only if include_total=true).
              example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
          description: Always false for error responses
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - UNAUTHORIZED
                - FORBIDDEN
                - PAYLOAD_TOO_LARGE
                - TOO_MANY_REQUESTS
                - REQUEST_TIMEOUT
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
                - DECIMAL_ERROR
                - UNSUPPORTED_MEDIA_TYPE
                - GATEWAY_TIMEOUT
                - NOT_ACCEPTABLE
                - PERSISTED_QUERY_INVALID
                - PERSISTED_QUERY_UNSUPPORTED_VERSION
                - PERSISTED_QUERY_INVALID_HASH
                - PERSISTED_QUERY_HASH_MISMATCH
                - PERSISTED_QUERY_NOT_FOUND
                - GRAPHQL_PARSE_ERROR
                - FEATURE_FLAG_DISABLED
                - CONFIGURATION_ERROR
                - RESTORE_ERROR
                - PERSISTENCE_ERROR
                - API_KEY_ERROR
                - DIAGNOSTICS_ERROR
                - SERVICE_SHUTTING_DOWN
                - HEALTH_CHECK_ERROR
                - CONSUMER_SUSPENDED
                - ENTRY_ALREADY_REPLAYED
                - INVALID_REQUEST
                - QUEUE_ERROR
                - INVALID_DELIVERY_ID
                - DELIVERY_NOT_FOUND
                - INVALID_PAGINATION
                - INVALID_OUTBOX_FILTER
                - INVALID_RETRY_REQUEST
                - DLQ_ITEM_NOT_FOUND
                - DLQ_PROCESS_ERROR
                - DLQ_RETRY_ERROR
                - INVALID_ENDPOINT_URL
                - OUTBOX_PROCESSING_ERROR
                - INVALID_CLEANUP_REQUEST
                - CLEANUP_ERROR
                - CORS_ORIGIN_DENIED
                - RATE_LIMIT_EXCEEDED
                - INVALID_ADDRESS
                - ERASURE_FAILED
                - INTERNAL_CALLER_ADDRESS_MISSING
                - unsupported_media_type
                - payload_too_large
                - invalid_encoding
                - payload_too_deep
                - invalid_json
                - missing_secret
                - missing_delivery_id
                - missing_timestamp
                - missing_signature
                - invalid_timestamp
                - timestamp_outside_tolerance
                - previous_secret_expired
                - signature_mismatch
                - duplicate_delivery
                - UNSUPPORTED_VERSION
                - IDEMPOTENCY_CONFLICT
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Validation failed
            requestId:
              type: string
              description: Correlation ID for log triage
              example: 550e8400-e29b-41d4-a716-446655440000
            details:
              description: Additional error context (field-level validation errors, etc.)
              nullable: true
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Include in Authorization header:
        Authorization: Bearer <token>
    AdminAuth:
      type: http
      scheme: bearer
      description: |
        Admin API key for privileged operations. Include in Authorization header:
        Authorization: Bearer <ADMIN_API_KEY>
        Set via ADMIN_API_KEY environment variable.
...